Move category and amenities filters to the "More filters" panel
